+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Current Master
 
 - Redesigned init, vault create, vault restore and seed confirmation screens.
+- Added pending transactions to transaction list on account screen.
+- Clicking a pending transaction takes you back to the transaction approval screen.
 
 ## 2.1.0 2016-05-26